+++ jest.config.js
@@ -4,6 +4,9 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
+
+const integrationTestPath = '<rootDir>/packages/test-integration';
+
 module.exports = {
     coverageThreshold: {
         global: {
@@ -13,4 +16,8 @@ module.exports = {
             statements: 80,
         },
     },
+    // Exclude integration tests from being run with config as it has its own config and setup
+    testPathIgnorePatterns: [integrationTestPath],
+    // Direct Jest to read the jest config file from integration tests
+    projects: ['<rootDir>', integrationTestPath],
 };