+++ packages/assert/src/assert.ts
@@ -5,10 +5,9 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-import axe, { ElementContext, RunOptions } from 'axe-core';
-import { extended } from '@sa11y/preset-rules';
-import { A11yConfig } from '@sa11y/preset-rules';
-import { Formatter, a11yResultsFormatter } from '@sa11y/format';
+import * as axe from 'axe-core';
+import { A11yConfig, extended } from '@sa11y/preset-rules';
+import { a11yResultsFormatter, Formatter } from '@sa11y/format';
 
 // Error message prefix for runtime exceptions when running axe
 // TODO (refactor): Should this be exported? Can private variables be imported in tests?
@@ -16,20 +15,27 @@ import { Formatter, a11yResultsFormatter } from '@sa11y/format';
 export const axeRuntimeExceptionMsgPrefix = 'Error running accessibility checks using axe:';
 
 /**
+ * Type def for context that can be checked for accessibility.
+ * Limiting to this subset from all options supported by axe for ease of use and maintenance.
+ */
+export type a11yCheckableContext = Document | HTMLElement;
+
+/**
  * Checks DOM for accessibility issues and throws an error if violations are found.
- * @param dom - DOM to be tested for accessibility
+ * @param context - DOM or HTMLElement to be tested for accessibility
  * @param rules - A11yConfig preset rule to use, defaults to extended
  * @param formatter - Function to format a11y violations. Passing null will format using JSON stringify.
- * @throws error - with the accessibility issues found
+ * @throws error - with the accessibility issues found, does not return any value
  * */
+// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
 export async function assertAccessible(
-    dom: Document = document,
+    context: a11yCheckableContext = document,
     rules: A11yConfig = extended,
     formatter: Formatter = a11yResultsFormatter
 ) {
     let violations;
     try {
-        const results = await axe.run(dom as ElementContext, rules as RunOptions);
+        const results = await axe.run(context as axe.ElementContext, rules as axe.RunOptions);
         violations = results.violations;
     } catch (e) {
         throw new Error(`${axeRuntimeExceptionMsgPrefix} ${e}`);