+++ packages/test-utils/src/index.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2020, salesforce.com, inc.
+ * All rights reserved.
+ * SPDX-License-Identifier: BSD-3-Clause
+ * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
+ */
+
+export { domWithA11yIssues, domWithNoA11yIssues, shadowDomID, domWithA11yIssuesBodyID } from './test-data';
+export { beforeEachSetup, cartesianProduct } from './utils';